Use int instead of size_t in Util::binarySearch

+++ core/Util.cpp
@@ -53,11 +53,11 @@ namespace core {
 // for the (possibly empty) remainder; Search returns the first true index. If there is no such
 // index, Search returns n. (Note that the "not found" return value is not -1 as in, for instance,
 // strings.Index). Search calls f(i) only for i in the range [0, n).
-size_t Util::binarySearch(size_t last, std::function<bool(size_t)> fn) {
-  size_t low = 0;
-  size_t high = last;
+int Util::binarySearch(int last, std::function<bool(int)> fn) {
+  int low = 0;
+  int high = last;
   while (low < high) {
-    size_t mid = (low + high) / 2;
+    int mid = (low + high) / 2;
     //    qDebug("low: %d, high: %d, mid: %d", low, high, mid);
     if (fn(mid)) {
       if (mid == 0) {
