+++ core/Util.h
@@ -19,7 +19,7 @@ class Util {
   DISABLE_COPY_AND_MOVE(Util)
 
  public:
-  static size_t binarySearch(size_t last, std::function<bool(size_t)> fn);
+  static int binarySearch(int last, std::function<bool(int)> fn);
   static void ensureDir(const QString& path);
 
   /**
